+++ frontend/components/monitoring/CreateSuiteModal.vue
@@ -77,3 +77,4 @@ const create = async () => {
 }
 </script>
 
+